# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Telemt Panel API
  version: 2.0.0-draft1
  description: >
    Admin API of Telemt Panel v2 plus the unauthenticated subscription page.
    Telemt-shaped payloads follow the upstream source catalog (specs/07-telemt-sdk.md).
servers:
  - url: /
    description: Same origin; base_path is prepended by the SPA runtime when configured.
...
security:
  - session: []
paths:
...
  /api/users/{username}/ip-history/reset:
    post:
      tags: [users]
      operationId: resetUserIPHistory
      description: Clears only address history, including pending writes. Later observations may create new history. Does not reset traffic or Telemt IP limits.
      parameters:
        - { $ref: "#/components/parameters/Username" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirm]
              properties:
                confirm: { type: boolean, const: true }
      responses:
        "204": { description: IP history cleared }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/users/{username}/traffic/reset:
    post:
      tags: [users]
      operationId: resetUserTraffic
      description: >
        Irreversibly removes the panel-observed total, current baseline and
        graph history for one account. The next coherent collector snapshot
        establishes a new baseline.
      parameters: [ { $ref: "#/components/parameters/Username" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DestructiveConfirmationRequest" }
      responses:
        "204": { description: User traffic reset }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: panel_session
  parameters:
    Username:
      name: username
      in: path
      required: true
      schema: { type: string, pattern: "^[A-Za-z0-9_.-]{1,64}$" }
  responses:
    InternalError:
      description: Internal panel error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    BadRequest:
      description: Invalid input
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - confirmation_required
            - invalid_credentials
            - rate_limited
            - session_expired
            - csrf_rejected
            - internal_error
            - not_found
            - telemt_unreachable
            - capability_absent
            - capability_unavailable
            - manual_restart_required
            - update_locked
            - sublink_unavailable
            - log_tail_unavailable
            - log_stream_unavailable
            - log_source_error
            - totp_required
            - invalid_totp
            - totp_already_enabled
            - totp_setup_changed
            - invalid_webauthn_origin
            - invalid_webauthn_challenge
            - invalid_webauthn_response
            - webauthn_credential_exists
            - passkey_unavailable
            - telemt_auth_failed
            - user_exists
            - last_user_forbidden
            - read_only
            - revision_conflict
            - invalid_toml
            - invalid_config_path
            - config_unset_unsupported
            - no_changes
            - toml_projection_failed
            - reload_in_progress
            - reload_not_found
            - ambiguous_listeners
            - access_not_editable
            - section_not_editable
            - field_not_editable
            - unauthorized
            - forbidden
            - method_not_allowed
            - config_patch_not_atomic
            - payload_too_large
            - api_disabled
            - maestro_unavailable
            - unsupported_media_type
            - web_runtime_unavailable
            - web_snapshot_busy
            - web_runtime_mismatch
            - web_issuance_enabled
            - web_operation_in_progress
            - web_session_not_found
            - web_operation_not_found
            - web_vhost_not_found
            - web_profile_required
          description: >
            Machine code. Panel codes actually emitted today (grepped from every
            WriteError call site): bad_request, invalid_credentials, rate_limited,
            session_expired, csrf_rejected, internal_error, not_found,
            telemt_unreachable, capability_absent, capability_unavailable,
            manual_restart_required, update_locked, sublink_unavailable,
            log_tail_unavailable, log_stream_unavailable, log_source_error,
            totp_required, invalid_totp, totp_already_enabled,
            totp_setup_changed, invalid_webauthn_origin,
            invalid_webauthn_challenge, invalid_webauthn_response,
            webauthn_credential_exists, passkey_unavailable,
            invalid_toml, invalid_config_path, config_unset_unsupported,
            no_changes, toml_projection_failed.
            capability_absent (501) vs capability_unavailable (503) are
            deliberately distinct, not aliases: capability_absent means the
            route itself doesn't exist on this Telemt build (a bare 404/405
            with no error envelope — detected reactively, after attempting
            the call: rotate-secret, enable/disable, POST /api/telemt/reload,
            GET /api/telemt/reload/{id}); capability_unavailable means the
            route exists but the feature behind it is switched off on this
            Telemt — either known up front from the SDK's cached Capabilities
            probe (GET/PATCH /api/telemt/config, config_api) or reported by
            the response itself (GET /api/telemt/tls-fingerprints, whose
            enabled:false means runtime_edge_enabled is off; read from the
            response rather than probed so an unreachable Telemt still maps
            to 502 telemt_unreachable).
            Reserved for milestones not yet implemented: telemt_auth_failed
            (superseded on /api/telemt/info by a reachable:false
            body, not an error status — kept here for /api/telemt/config, M3).
            A well-formed Telemt *APIError whose status is 4xx and isn't otherwise
            mapped above is passed through verbatim with Telemt's own code — notably
            user_exists, last_user_forbidden, read_only, revision_conflict,
            reload_in_progress, reload_not_found, ambiguous_listeners (the latter two
            absent from Telemt's own documented error-code table but confirmed against
            its source, M3), plus any other code in Telemt's own set (07-telemt-sdk.md):
            bad_request, access_not_editable, section_not_editable, field_not_editable,
            unauthorized, forbidden, method_not_allowed, config_patch_not_atomic,
            payload_too_large, api_disabled, maestro_unavailable — except
            access_not_editable/section_not_editable/field_not_editable/
            config_patch_not_atomic/ambiguous_listeners on PATCH /api/telemt/config,
            which the panel remaps to HTTP 422 regardless of Telemt's own status.
            The WEB group (Telemt >= 3.5.3, internal/telemt/types_web.go) adds
            web_runtime_mismatch, web_issuance_enabled, web_operation_in_progress,
            web_snapshot_busy, web_session_not_found, web_operation_not_found and
            unsupported_media_type; web_runtime_unavailable is listed because it is
            Telemt's own code, but the panel remaps it to capability_unavailable
            (rule R5) so the closed-capability gate is drawn instead of an error.
            Every code in this enum must carry a message in BOTH dictionaries —
            web/src/i18n/i18n.test.ts walks this list.
        message: { type: string }
...
    DestructiveConfirmationRequest:
      type: object
      additionalProperties: false
      required: [confirm]
      properties:
        confirm: { type: boolean, const: true }
...